Crate stahl_errors

A common error type for Stahl.

This may be generic enough to split out for more errors; just about whenever Location is relevant, this is probably the error type to use.

Macros

ensure
The equivalent of assert!, but using raise! instead of panic!.

ensure_eq
The equivalent of assert_eq!, but using raise! instead of panic!.

ensure_ne
The equivalent of assert_ne!, but using raise! instead of panic!.

err
Evaluates to an Error constructed from a formatting expression.

raise
Returns an Error constructed from a formatting expression.

todo
An equivalent of unimplemented!() that returns an Error instead.

Structs

Error
The common error wrapper, which allows attaching a location and/or a cause.

Location
The location at which an error occurred.

PointLC
A point in source code. A (byte, line, column) triple.

Enums

Position
The point or span at which an error occurred.

Traits

ResultExt
An extension trait for Results.

Type Definitions

Result
A useful alias for Result with this crate's Error type.